+++ src/gort/devices/spec.py
@@ -13,7 +13,7 @@ import asyncio
 from typing import TYPE_CHECKING
 
 from gort.exceptions import ErrorCodes, GortError, GortSpecError
-from gort.exposure import Exposure
+from gort.exposure import UPDATE_HEADER_CB_TYPE, Exposure
 from gort.gort import GortDevice, GortDeviceSet
 from gort.recipes.calibration import CalibrationRecipe
 from gort.tools import is_interactive, is_notebook
@@ -341,6 +341,7 @@ class SpectrographSet(GortDeviceSet[Spectrograph]):
         async_readout: bool = False,
         count: int = 1,
         object: str | None = None,
+        update_header_cb: UPDATE_HEADER_CB_TYPE = None,
     ) -> Exposure | list[Exposure]:
         """Exposes the spectrographs.
 
@@ -367,6 +368,10 @@ class SpectrographSet(GortDeviceSet[Spectrograph]):
         object
             A string that will be stored in the ``OBJECT`` header
             keyword.
+        update_header_cb
+            A function that will be called after integration but before
+            readout. It receives the current header and must modify it
+            in place.
 
         Returns
         -------
@@ -428,6 +433,7 @@ class SpectrographSet(GortDeviceSet[Spectrograph]):
 
             exposure = Exposure(seqno, self, flavour=flavour)
             exposure.object = object or ""
+            exposure._update_header_cb = update_header_cb
 
             await exposure.expose(
                 exposure_time=exposure_time,