+++ src/gort/observer.py
@@ -13,11 +13,15 @@ import logging
 import re
 from time import time
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
+
+import numpy
+import pandas
+from astropy.time import Time
 
 from gort.exceptions import GortObserverError
 from gort.tile import Coordinates
-from gort.tools import cancel_task
+from gort.tools import build_guider_reply_list, cancel_task
 from gort.transforms import fibre_slew_coordinates
 
 
@@ -51,6 +55,10 @@ class GortObserver:
         self.mask_positions = self._get_mask_positions(mask_positions_pattern)
 
         self.guide_task: asyncio.Future | None = None
+        self.guider_monitor = GuiderMonitor(self)
+
+        self.has_standards: bool = True
+        self.standards = Standards(self, tile, self.mask_positions)
 
     def __repr__(self):
         return f"<GortObserver (tile_id={self.tile.tile_id})>"
@@ -58,20 +66,22 @@ class GortObserver:
     async def slew(self):
         """Slew to the telescope fields."""
 
+        tile = self.tile
+
         cotasks = []
 
         # Stops guiders.
         await self.gort.guiders.stop()
 
         # Slew telescopes.
-        self.write_to_log(f"Slewing to tile_id={self.tile.tile_id}.", level="info")
+        self.write_to_log(f"Slewing to tile_id={tile.tile_id}.", level="info")
 
-        sci = (self.tile.sci_coords.ra, self.tile.sci_coords.dec)
-        self.write_to_log(f"Science: {str(self.tile.sci_coords)}")
+        sci = (tile.sci_coords.ra, tile.sci_coords.dec, tile.sci_coords.pa)
+        self.write_to_log(f"Science: {str(tile.sci_coords)}")
 
         spec = None
-        if self.tile.spec_coords and len(self.tile.spec_coords) > 0:
-            first_spec = self.tile.spec_coords[0]
+        if tile.spec_coords and len(tile.spec_coords) > 0:
+            first_spec = tile.spec_coords[0]
 
             # For spec we slew to the fibre with which we'll observe first.
             # This should save a bit of time converging.
@@ -86,8 +96,8 @@ class GortObserver:
 
         sky = {}
         for skytel in ["SkyE", "SkyW"]:
-            if skytel.lower() in self.tile.sky_coords:
-                sky_coords_tel = self.tile.sky_coords[skytel.lower()]
+            if skytel.lower() in tile.sky_coords:
+                sky_coords_tel = tile.sky_coords[skytel.lower()]
                 if sky_coords_tel is not None:
                     sky[skytel.lower()] = (sky_coords_tel.ra, sky_coords_tel.dec)
                     self.write_to_log(f"{skytel}: {sky_coords_tel}")
@@ -165,12 +175,14 @@ class GortObserver:
 
         if "spec" not in guide_on_telescopes:
             self.write_to_log("No standards defined. Blocking fibre mask.", "warning")
+            self.has_standards = False
             await self.gort.telescopes.spec.fibsel.move_relative(500)
 
         if n_skies == 0:
             self.write_to_log("No sky positions defined.", "warning")
 
         # Start guide loop.
+        await self.guider_monitor.restart()
         self.guide_task = asyncio.gather(*guide_coros)
 
         await asyncio.sleep(5)
@@ -217,7 +229,6 @@ class GortObserver:
         self,
         exposure_time: float = 900.0,
         show_progress: bool | None = None,
-        iterate_over_standards: bool = True,
         count: int = 1,
         **kwargs,
     ):
@@ -229,9 +240,6 @@ class GortObserver:
             The length of the exposure in seconds.
         show_progress
             Displays a progress bar with the elapsed exposure time.
-        iterate_over_standards
-            Whether to move the spec telescope during intergration
-            to observe various standard stars in different fibres.
         count
             Number of exposures. If ``iterate_over_standards=True``, a
             full sequence of standards will be observed during each
@@ -247,8 +255,6 @@ class GortObserver:
         if "object" not in kwargs:
             kwargs["object"] = self.tile.object
 
-        header = await self._get_header()
-
         exposures: list[Exposure] = []
         standard_task: asyncio.Task | None = None
 
@@ -258,16 +264,20 @@ class GortObserver:
                 "info",
             )
 
-            if iterate_over_standards:
-                standard_task = asyncio.create_task(
-                    self._iterate_over_mask_positions(exposure_time)
-                )
+            # Refresh guider data for this exposure.
+            if nexp > 1:
+                await self.guider_monitor.restart()
+
+            # Move fibre selector to the first position. Should be there unless
+            # count > 1 in which case we need to move it back.
+            if self.has_standards:
+                await self.standards.start_iterating(exposure_time)
 
             exposure = await self.gort.specs.expose(
                 exposure_time=exposure_time,
-                header=header,
                 show_progress=show_progress,
                 count=1,
+                update_header_cb=self._update_header,
                 **kwargs,
             )
             assert not isinstance(exposure, list)
@@ -330,7 +340,195 @@ class GortObserver:
 
         return sorted(positions, key=lambda p: mask_config[p])
 
-    async def _iterate_over_mask_positions(self, exposure_time: float):
+    async def _update_header(self, header: dict[str, Any]):
+        """Updates the exposure header with pointing and guiding information."""
+
+        tile_id = self.tile.tile_id
+        dither_pos = self.tile.dither_position
+
+        tile_header = {
+            "TILE_ID": (tile_id or -999, "The tile_id of this observation"),
+            "DPOS": (dither_pos, "Dither position"),
+            "POSCIRA": round(self.tile.sci_coords.ra, 6),
+            "POSCIDE": round(self.tile.sci_coords.dec, 6),
+        }
+
+        if self.tile["skye"]:
+            tile_header.update(
+                {
+                    "POSKYERA": round(self.tile.sky_coords["skye"].ra, 6),
+                    "POSKYEDE": round(self.tile.sky_coords["skye"].dec, 6),
+                }
+            )
+
+        if self.tile["skyw"]:
+            tile_header.update(
+                {
+                    "POSKYWRA": round(self.tile.sky_coords["skyw"].ra, 6),
+                    "POSKYWDE": round(self.tile.sky_coords["skyw"].dec, 6),
+                }
+            )
+
+        header.update(tile_header)
+
+        self.guider_monitor.update_data()
+        header.update(self.guider_monitor.to_header())
+
+        # At this point the shutter is closed so let's stop observing standards.
+        # This also finishes updating the standards table.
+        if self.has_standards:
+            await self.standards.cancel()
+            header.update(self.standards.to_header())
+
+
+class GuiderMonitor:
+    """Monitors guider exposures."""
+
+    def __init__(self, observer: GortObserver):
+        self.observer = observer
+        self.gort = observer.gort
+
+        self.guider_task: asyncio.Task | None = None
+        self.guider_data: pandas.DataFrame | None = None
+
+        self._current_data = []
+
+    async def start(self):
+        """Starts monitoring."""
+
+        self.guider_task = asyncio.create_task(self._guider_monitor())
+
+    async def stop(self):
+        """Stops monitoring."""
+
+        await cancel_task(self.guider_task)
+
+    async def restart(self):
+        """Restarts monitoring."""
+
+        self.guider_data = None
+
+        await self.stop()
+        await self.start()
+
+    async def _guider_monitor(self):
+        """Monitors the guider data and build a data frame."""
+
+        self._current_data = []
+
+        task = asyncio.create_task(
+            build_guider_reply_list(
+                self.gort,
+                self._current_data,
+            )
+        )
+
+        try:
+            while True:
+                # Just keep the task running.
+                await asyncio.sleep(1)
+
+        except asyncio.CancelledError:
+            await cancel_task(task)
+            self.update_data()
+
+    def update_data(self):
+        """Updates the guider data frame."""
+
+        if len(self._current_data) > 0:
+            # Build DF with all the frames.
+            df = pandas.DataFrame.from_records(self._current_data)
+
+            # Group by frameno, keep only non-NaN values.
+            df = df.groupby(["frameno", "telescope"], as_index=False).apply(
+                lambda g: g.fillna(method="bfill", axis=0).iloc[0, :]
+            )
+
+            # Remove NaN rows.
+            df = df.dropna()
+
+            # Sort by frameno.
+            df = df.sort_values("frameno")
+
+            self.guider_data = df
+
+    def to_header(self):
+        """Returns a header with pointing and guiding information."""
+
+        header: dict[str, Any] = {}
+
+        if self.guider_data is not None:
+            for tel in ["sci", "spec", "skye", "skyw"]:
+                try:
+                    tel_data = self.guider_data.loc[self.guider_data.telescope == tel]
+                    if len(tel_data) < 2:
+                        frame0 = None
+                        framen = None
+                    else:
+                        frame0 = int(tel_data.frameno.min())
+                        framen = int(tel_data.frameno.max())
+
+                    header.update(
+                        {
+                            f"G{tel.upper()}FR0": (frame0, f"{tel} first guider frame"),
+                            f"G{tel.upper()}FRN": (framen, f"{tel} last guider frame"),
+                        }
+                    )
+
+                except Exception as err:
+                    self.gort.specs.write_to_log(
+                        f"Failed updating guider header information for {tel}: {err}",
+                        "warning",
+                    )
+                    continue
+
+        return header
+
+
+class Standards:
+    """Iterates over standards and monitors observed standards."""
+
+    def __init__(self, observer: GortObserver, tile: Tile, mask_positions: list[str]):
+        self.observer = observer
+        self.gort = observer.gort
+
+        self.tile = tile
+        self.mask_positions = mask_positions
+
+        self.iterate_task: asyncio.Task | None = None
+
+        self.current_standard: int = 1
+        self.standards = self._get_frame()
+
+    def _get_frame(self):
+        """Constructs the standard data frame."""
+
+        stdn = list(range(1, len(self.tile.spec_coords) + 1))
+        source_id = [cc.source_id or -1 for cc in self.tile.spec_coords]
+        ra = [cc.ra for cc in self.tile.spec_coords]
+        dec = [cc.dec for cc in self.tile.spec_coords]
+
+        default = [0] * len(stdn)
+
+        df = pandas.DataFrame(
+            {
+                "n": pandas.Series(stdn, dtype=numpy.int16),
+                "source_id": pandas.Series(source_id, dtype=numpy.int64),
+                "ra": pandas.Series(ra, dtype=numpy.float64),
+                "dec": pandas.Series(dec, dtype=numpy.float64),
+                "acquired": pandas.Series(default, dtype=numpy.int16),
+                "observed": pandas.Series(default, dtype=numpy.int16),
+                "t0": pandas.Series(default, dtype=numpy.float64),
+                "t1": pandas.Series(default, dtype=numpy.float64),
+                "fibre": pandas.Series([""] * len(stdn), dtype=str),
+            }
+        )
+
+        df.set_index("n", inplace=True)
+
+        return df
+
+    async def start_iterating(self, exposure_time: float):
         """Iterates over the fibre mask positions.
 
         Moving the ``spec`` telescope to each one of the standard star,
@@ -338,8 +536,29 @@ class GortObserver:
 
         """
 
-        # TODO: record how long we exposed on each standard and save that
-        # information somewhere.
+        await self.gort.telescopes.spec.fibsel.move_to_position(self.mask_positions[0])
+        await self.cancel()
+
+        self.standards = self._get_frame()
+        self.current_standard = 1
+
+        self.standards.loc[1, "acquired"] = 1
+        self.standards.loc[1, "t0"] = time()
+        self.standards.loc[1, "fibre"] = self.mask_positions[0]
+
+        self.iterate_task = asyncio.create_task(self._iterate(exposure_time))
+
+    async def cancel(self):
+        """Cancels iteration."""
+
+        await cancel_task(self.iterate_task)
+
+        if self.standards.loc[self.current_standard].acquired == 1:
+            self.standards.loc[self.current_standard, "observed"] = 1
+            self.standards.loc[self.current_standard, "t1"] = time()
+
+    async def _iterate(self, exposure_time: float):
+        """Iterate task."""
 
         # Time to acquire a standard.
         ACQ_PER_STD = 30
@@ -364,9 +583,10 @@ class GortObserver:
         # how long we expect to take acquiring.
         time_per_position = exposure_time / n_stds - ACQ_PER_STD
         if time_per_position < 0:
-            self.write_to_log(
+            self.observer.write_to_log(
                 "Exposure time is too short to observe this "
-                "many standards. I will do what I can."
+                "many standards. I will do what I can.",
+                "warning",
             )
             time_per_position = exposure_time / n_stds
 
@@ -386,8 +606,11 @@ class GortObserver:
             # exposure there is no point in going to the next standard.
             t_now = time()
             if t_now - t0 > exposure_time - 2 * ACQ_PER_STD:
-                self.write_to_log("Exiting standard loop.")
-                self.write_to_log(f"Standards observed: {n_observed}/{n_stds}.", "info")
+                self.observer.write_to_log("Exiting standard loop.")
+                self.observer.write_to_log(
+                    f"Standards observed: {n_observed}/{n_stds}.",
+                    "info",
+                )
                 return
 
             # Time to move to another standard?
@@ -397,8 +620,17 @@ class GortObserver:
                 if len(spec_coords) == current_std_idx + 1:
                     continue
 
+                # Moving the mask to an intermediate position while we move around.
+                spec_tel = self.gort.telescopes.spec
+                await spec_tel.fibsel.move_relative(500)
+
+                # Register the previous standard.
+                self.standards.loc[self.current_standard, "t1"] = time()
+                self.standards.loc[self.current_standard, "observed"] = True
+
                 # Increase current index and get coordinates.
                 current_std_idx += 1
+                self.current_standard += 1
 
                 # New coordinates to observe.
                 new_coords = spec_coords[current_std_idx]
@@ -411,16 +643,11 @@ class GortObserver:
                 # precise metrology.
                 new_guider_pixel = guider_pixels[new_mask_position]
 
-                self.write_to_log(
+                self.observer.write_to_log(
                     f"Moving to standard #{current_std_idx+1} ({new_coords}) "
                     f"on fibre {new_mask_position}.",
                     "info",
                 )
-
-                spec_tel = self.gort.telescopes.spec
-
-                # Moving the mask to an intermediate position while we move around.
-                await spec_tel.fibsel.move_relative(500)
 
                 # Slew to new coordinates. We actually slew to the coordinates
                 # that make the new star close to the fibre that will observe it.
@@ -432,7 +659,7 @@ class GortObserver:
                 )
 
                 # Finish guiding on spec telescope.
-                self.write_to_log("Stopping guiding on spec and reslewing.")
+                self.observer.write_to_log("Stopping guiding on spec and reslewing.")
 
                 cotasks = [
                     self.gort.guiders["spec"].stop(),
@@ -443,7 +670,7 @@ class GortObserver:
                 # Start to guide. Note that here we use the original coordinates
                 # of the star along with the pixel on the master frame on which to
                 # guide. See the note in fibre_slew_coordinates().
-                self.write_to_log("Starting to guide on spec telescope.")
+                self.observer.write_to_log("Starting to guide on spec telescope.")
                 asyncio.create_task(
                     self.gort.guiders.spec.guide(
                         ra=new_coords.ra,
@@ -455,13 +682,13 @@ class GortObserver:
 
                 result = await self.gort.guiders.spec.wait_until_guiding(timeout=60)
                 if result[0] is False:
-                    self.write_to_log(
+                    self.observer.write_to_log(
                         "Failed to acquire standard position. Skipping.",
                         "warning",
                     )
                     continue
 
-                self.write_to_log(f"Standard position {new_mask_position} acquired.")
+                self.observer.write_to_log(f"Standard {new_mask_position} acquired.")
 
                 # Move mask to uncover fibre.
                 await spec_tel.fibsel.move_to_position(new_mask_position)
@@ -469,33 +696,25 @@ class GortObserver:
                 n_observed += 1
                 t0_last_std = time()
 
-    async def _get_header(self):
-        """Returns the extra header dictionary from the tile data."""
+                self.standards.loc[self.current_standard, "acquired"] = 1
+                self.standards.loc[self.current_standard, "t0"] = time()
+                self.standards.loc[self.current_standard, "fibre"] = new_mask_position
 
-        tile_id = self.tile.tile_id
-        dither_pos = self.tile.dither_position
+    def to_header(self):
+        """Returns observed standards as a header-ready dictionary."""
 
-        header = {
-            "tile_id": (tile_id or -999, "The tile_id of this observation"),
-            "dpos": (dither_pos, "Dither position"),
-            "poscira": round(self.tile.sci_coords.ra, 6),
-            "poscide": round(self.tile.sci_coords.dec, 6),
-        }
+        header_data = {}
 
-        if self.tile["skye"]:
-            header.update(
-                {
-                    "poskyera": round(self.tile.sky_coords["skye"].ra, 6),
-                    "poskyede": round(self.tile.sky_coords["skye"].dec, 6),
-                }
-            )
+        for nstd, data in self.standards.iterrows():
+            header_data[f"STD{nstd}ID"] = data.source_id if data.source_id > 0 else None
+            header_data[f"STD{nstd}RA"] = data.ra
+            header_data[f"STD{nstd}DE"] = data.dec
+            header_data[f"STD{nstd}ACQ"] = bool(data.observed)
 
-        if self.tile["skyw"]:
-            header.update(
-                {
-                    "poskywra": round(self.tile.sky_coords["skyw"].ra, 6),
-                    "poskywde": round(self.tile.sky_coords["skyw"].dec, 6),
-                }
-            )
+            if data.observed:
+                header_data[f"STD{nstd}T0"] = Time(data.t0, format="unix").isot
+                header_data[f"STD{nstd}T1"] = Time(data.t1, format="unix").isot
+                header_data[f"STD{nstd}EXP"] = round(data.t1 - data.t0, 1)
+                header_data[f"STD{nstd}FIB"] = data.fibre
 
-        return header.copy()
+        return header_data