+++ src/gort/tile.py
@@ -17,6 +17,7 @@ from astropy.coordinates import EarthLocation, SkyCoord
 from astropy.time import Time
 from httpx import RequestError
 
+from gort import config
 from gort.exceptions import GortNotImplemented, GortWarning, TileError
 from gort.tools import get_calibrators_sync, get_db_connection, get_next_tile_id_sync
 from gort.transforms import fibre_to_master_frame
@@ -44,6 +45,8 @@ class Coordinates:
         The RA coordinate, in degrees. FK5 frame at the epoch of observation.
     dec
         The Dec coordinate, in degrees.
+    pa
+        Position angle of the IFU. Defaults to PA=0.
     centre_on_fibre
         The name of the fibre on which to centre the target, with the format
         ``<ifulabel>-<finufu>``. By default, acquires the target on the central
@@ -51,9 +54,17 @@ class Coordinates:
 
     """
 
-    def __init__(self, ra: float, dec: float, centre_on_fibre: str | None = None):
+    def __init__(
+        self,
+        ra: float,
+        dec: float,
+        pa: float | None = None,
+        centre_on_fibre: str | None = None,
+    ):
         self.ra = ra
         self.dec = dec
+        self.pa = pa if pa is not None else 0.0
+
         self.skycoord = SkyCoord(ra=ra, dec=dec, unit="deg", frame="fk5")
 
         self.centre_on_fibre = centre_on_fibre
@@ -62,17 +73,21 @@ class Coordinates:
         self._mf_pixel = self.set_mf_pixel(centre_on_fibre)
 
     def __repr__(self):
-        return f"<{self.__class__.__name__} (ra={self.ra:.6f}, dec={self.dec:.6f})>"
+        return (
+            f"<{self.__class__.__name__} "
+            f"(ra={self.ra:.6f}, dec={self.dec:.6f}, pa={self.pa:.3f})>"
+        )
 
     def __str__(self):
-        return f"{self.ra:.6f}, {self.dec:.6f}"
+        return f"{self.ra:.6f}, {self.dec:.6f}, {self.pa:.3f}"
 
     def calculate_altitude(self, time: Time | None = None):
         """Returns the current altitude of the target."""
 
         if time is None:
             time = Time.now()
-        location = EarthLocation.of_site("Las Campanas Observatory")
+
+        location = EarthLocation.from_geodetic(**config["site"])
 
         sc = self.skycoord.copy()
         sc.obstime = time
@@ -172,7 +187,8 @@ class QuerableCoordinates(Coordinates):
 
         # Exclude targets that are too low.
         if exclude_invisible:
-            skycoords.location = EarthLocation.of_site("Las Campanas Observatory")
+            skycoords.location = EarthLocation.from_geodetic(**config["site"])
+
             skycoords.obstime = Time.now()
             altaz_skycoords = skycoords.transform_to("altaz")
             skycoords = skycoords[altaz_skycoords.alt.deg > 30]
@@ -215,6 +231,13 @@ class ScienceCoordinates(Coordinates):
         The RA coordinate, in degrees. FK5 frame at the epoch of observation.
     dec
         The Dec coordinate, in degrees.
+    pa
+        Position angle of the IFU. Defaults to PA=0.
+    centre_on_fibre
+        The name of the fibre on which to centre the target, with the format
+        ``<ifulabel>-<finufu>``. By default, acquires the target on the central
+        fibre of the science IFU.
+
     """
 
 
@@ -225,9 +248,19 @@ class SkyCoordinates(QuerableCoordinates):
 
 
 class StandardCoordinates(QuerableCoordinates):
-    """A standard position."""
+    """A standard position.
+
+    In addition to the `.QuerableCoordinates` arguments the class acceps
+    a ``source_id`` Gaia identifier.
+
+    """
 
     __db_table__ = "lvmopsdb.standard"
+
+    def __init__(self, *args, source_id: int | None = None, **kwargs):
+        self.source_id = source_id
+
+        super().__init__(*args, **kwargs)
 
 
 class Tile(dict[str, Coordinates | list[Coordinates] | None]):
@@ -280,8 +313,11 @@ class Tile(dict[str, Coordinates | list[Coordinates] | None]):
         return (
             "<Tile "
             f"(tile_id={self.tile_id}, "
-            f"science ra={self.sci_coords.ra:.6f}, dec={self.sci_coords.dec:.6f}; "
-            f"n_skies={len(self.sky_coords)}; n_standards={len(self.spec_coords)})>"
+            f"science ra={self.sci_coords.ra:.6f}, "
+            f"dec={self.sci_coords.dec:.6f}, "
+            f"pa={self.sci_coords.pa:.3f}; "
+            f"n_skies={len(self.sky_coords)}; "
+            f"n_standards={len(self.spec_coords)})>"
         )
 
     @property
@@ -342,6 +378,7 @@ class Tile(dict[str, Coordinates | list[Coordinates] | None]):
         cls,
         ra: float,
         dec: float,
+        pa: float = 0.0,
         sky_coords: dict[str, SkyCoordinates | CoordTuple] | None = None,
         spec_coords: list[StandardCoordinates | CoordTuple] | None = None,
         **kwargs,
@@ -350,8 +387,10 @@ class Tile(dict[str, Coordinates | list[Coordinates] | None]):
 
         Parameters
         ----------
-        sci_coords
+        ra,dec
             The science telescope pointing.
+        pa
+            Position angle of the science IFU. Defaults to PA=0.
         sky_coords
             A dictionary of ``skye`` and ``skyw`` coordinates. If `None`,
             autocompleted from the closest available regions.
@@ -363,29 +402,27 @@ class Tile(dict[str, Coordinates | list[Coordinates] | None]):
 
         """
 
-        sci_coords = ScienceCoordinates(ra, dec)
+        sci_coords = ScienceCoordinates(ra, dec, pa=pa)
+
+        calibrators: dict | None = None
+        if sky_coords is None or spec_coords is None:
+            calibrators = get_calibrators_sync(ra=ra, dec=dec)
 
         if sky_coords is None:
-            exclude_coordinates: list[CoordTuple] = []
+            assert calibrators is not None
             sky_coords = {}
-            for telescope in ["skye", "skyw"]:
-                coords = SkyCoordinates.from_science_coordinates(
-                    sci_coords,
-                    exclude_coordinates=exclude_coordinates,
-                )
-                sky_coords[telescope] = coords
-                exclude_coordinates.append((coords.ra, coords.dec))
+            sky_coords["skye"] = SkyCoordinates(*calibrators["sky_pos"][0])
+            sky_coords["skyw"] = SkyCoordinates(*calibrators["sky_pos"][1])
 
         if spec_coords is None:
-            exclude_coordinates: list[CoordTuple] = []
+            assert calibrators is not None
             spec_coords = []
-            for _ in range(12):
-                coords = StandardCoordinates.from_science_coordinates(
-                    sci_coords,
-                    exclude_coordinates=exclude_coordinates,
+            for ii in range(12):
+                coords = StandardCoordinates(
+                    *calibrators["standard_pos"][ii],
+                    source_id=calibrators["standard_ids"][ii],
                 )
                 spec_coords.append(coords)
-                exclude_coordinates.append((coords.ra, coords.dec))
 
         return cls(
             sci_coords,
@@ -401,6 +438,7 @@ class Tile(dict[str, Coordinates | list[Coordinates] | None]):
         tile_id: int | None = None,
         ra: float | None = None,
         dec: float | None = None,
+        pa: float = 0.0,
         **kwargs,
     ):
         """Creates a new instance of :obj:`.Tile` with data from the scheduler.
@@ -416,6 +454,8 @@ class Tile(dict[str, Coordinates | list[Coordinates] | None]):
             Calibrators will be selected from the scheduler.
         dec
             Declination coordinates of the science telescope pointing.
+        pa
+            Position angle of the science IFU. Defaults to PA=0.
         kwargs
             Arguments to be passed to the initialiser.
 
@@ -428,15 +468,15 @@ class Tile(dict[str, Coordinates | list[Coordinates] | None]):
                 raise TileError("Cannot retrieve tile_id from scheduler.")
 
             tile_id = tile_id_data["tile_id"]
-            sci_pos = tile_id_data["tile_pos"][:2]
-            dither_pos = 0
+            sci_pos = tile_id_data["tile_pos"]
+            dither_pos = tile_id_data["dither_pos"]
 
         elif tile_id is not None:
             raise GortNotImplemented("Initialising from a tile_id is not supported.")
 
-        elif tile_id is None and (ra is not None and dec is not None):
+        elif not tile_id and (ra is not None and dec is not None):
             tile_id = None
-            sci_pos = (ra, dec)
+            sci_pos = (ra, dec, pa)
             dither_pos = 0
 
         else:
@@ -444,16 +484,24 @@ class Tile(dict[str, Coordinates | list[Coordinates] | None]):
 
         sci_coords = ScienceCoordinates(*sci_pos, centre_on_fibre=None)
 
-        calibrator_data = get_calibrators_sync(
-            tile_id=None,
-            ra=sci_pos[0],
-            dec=sci_pos[1],
-        )
+        if tile_id:
+            calibrator_data = get_calibrators_sync(tile_id=tile_id)
+        else:
+            calibrator_data = get_calibrators_sync(ra=sci_pos[0], dec=sci_pos[1])
+
         sky_coords = {
             "skye": calibrator_data["sky_pos"][0],
             "skyw": calibrator_data["sky_pos"][1],
         }
-        spec_coords = list(calibrator_data["standard_pos"])
+
+        spec_coords = []
+        for ii in range(len(calibrator_data["standard_pos"])):
+            spec_coords.append(
+                StandardCoordinates(
+                    *calibrator_data["standard_pos"][ii],
+                    source_id=calibrator_data["standard_ids"][ii],
+                )
+            )
 
         new_obj = cls(
             sci_coords,